Move bulk edit views to shared location
