+++ src/js/views/patients/shared/bulk-edit/bulk-edit-action-body.hbs
@@ -1,0 +1,9 @@
+<div class="flex u-margin--t-16"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.shared.bulkEdit.bulkEditViews.bulkEditActionBodyTemplate.stateLabel }}</h4><div class="flex-grow" data-state-region></div></div>
+<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.shared.bulkEdit.bulkEditViews.bulkEditActionBodyTemplate.ownerLabel }}</h4><div class="flex-grow" data-owner-region></div></div>
+<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.shared.bulkEdit.bulkEditViews.bulkEditActionBodyTemplate.dueDateLabel }}</h4>
+  <div class="flex flex-grow">{{~ remove_whitespace ~}}
+    <div class="flex-grow" data-due-date-region></div>{{~ remove_whitespace ~}}
+    <div class="flex-grow" data-due-time-region></div>{{~ remove_whitespace ~}}
+  </div>
+</div>
+<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.shared.bulkEdit.bulkEditViews.bulkEditActionBodyTemplate.durationLabel }}</h4><div class="flex-grow" data-duration-region></div></div>
