+++ src/js/views/patients/shared/bulk-edit/bulk-edit-flow-body.hbs
@@ -1,0 +1,2 @@
+<div class="flex u-margin--t-16"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.shared.bulkEdit.bulkEditViews.bulkEditFlowBodyTemplate.stateLabel }}</h4><div class="flex-grow" data-state-region></div></div>
+<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.shared.bulkEdit.bulkEditViews.bulkEditFlowBodyTemplate.ownerLabel }}</h4><div class="flex-grow" data-owner-region></div></div>
