--- src/js/views/patients/worklist/bulk-edit/bulk-edit-action-body.hbs
@@ -1,9 +1,0 @@
-<div class="flex u-margin--t-16"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.worklist.bulkEditViews.bulkEditActionBodyTemplate.stateLabel }}</h4><div class="flex-grow" data-state-region></div></div>
-<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.worklist.bulkEditViews.bulkEditActionBodyTemplate.ownerLabel }}</h4><div class="flex-grow" data-owner-region></div></div>
-<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.worklist.bulkEditViews.bulkEditActionBodyTemplate.dueDateLabel }}</h4>
-  <div class="flex flex-grow">{{~ remove_whitespace ~}}
-    <div class="flex-grow" data-due-date-region></div>{{~ remove_whitespace ~}}
-    <div class="flex-grow" data-due-time-region></div>{{~ remove_whitespace ~}}
-  </div>
-</div>
-<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.worklist.bulkEditViews.bulkEditActionBodyTemplate.durationLabel }}</h4><div class="flex-grow" data-duration-region></div></div>
