--- src/js/views/patients/worklist/bulk-edit/bulk-edit-flow-body.hbs
@@ -1,2 +1,0 @@
-<div class="flex u-margin--t-16"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.worklist.bulkEditViews.bulkEditFlowBodyTemplate.stateLabel }}</h4><div class="flex-grow" data-state-region></div></div>
-<div class="flex u-margin--t-8"><h4 class="sidebar__label u-margin--t-8">{{ @intl.patients.worklist.bulkEditViews.bulkEditFlowBodyTemplate.ownerLabel }}</h4><div class="flex-grow" data-owner-region></div></div>
