--- src/js/views/patients/worklist/bulk-edit/bulk-edit.scss
@@ -1,7 +1,0 @@
-.bulk-edit {
-  min-width: 464px;
-}
-
-.bulk-edit__body {
-  padding: 0 24px;
-}
