+++ src/js/views/patients/worklist/worklist-list.scss
@@ -96,15 +96,6 @@
   margin-right: 16px;
 }
 
-.worklist-list__bulk-edit-cancel {
-  color: $black-60;
-  cursor: pointer;
-  font-size: 14px;
-  font-weight: 600px;
-  margin-left: 16px;
-  vertical-align: middle;
-}
-
 .worklist-list__flow-status {
   border: 1px solid transparent;
   display: inline-block;
